This commit adds in a bridge between the audio pipeline and native audio client.
The rationale is as follows:

Once a MsgDecoded message is received we know the specifics of the stream.
At this point we check that the native audio engine can play the stream.

If so we proceed to initialise the native audio and create a native
render thread.

The native render thread attempts to pre-populate the native render buffer
before starting the native audio client.

As pipeline data is read is put in a FIFO. The native render thread then
reads the data from the FIFO when prompted by the audio client.