Add Pipeline Observer to SampleMediaPlayer to handle playback control
We need to properly control playback state for the pipeline so add a PipelineObserver
to the SampleMediaPlayer which will manage notifications about pipeline
state changes, and also record whether the current stream is a live stream, in
which case we will inhibit requests to 'pause' as it doesn't make sense.

We also remove the logging pipeline observer from the BaseMediaPlayer.