This commit adds support for audio session disconnections.
If the user modifies the configuration or state of the multimedia device
while we are running this needs to be dealt with.

To do this we implement the IAudioSessionEvents interface in
AudioSessionEvents.cpp and register it via RegisterAudioSessionNotification().

The OnSessionDisconnected() callback will be called when our connection
to the endpoint is lost.

Here we:

1. Send the _DisconnectedEvent to the PipelineAnimator thread.

   This will unblock the WaitForMultipleObjects() call. From this point
   on we pull from the pipeline, without waiting on the native renderer as
   it will have stopped.

   Audio data will be discarded. We are really waiting just waiting for a
   quit message.

2. Send a WM_APP_AUDIO_DISCONNECTED message to the UI thread.

   The UI thread will display a warning, instructing the user to restart
   the application, in the system tray, and terminate the media player.