Use the registry for persistent data storage
The original implementation for persistent storage used the Windows Property
System, adding custom properties to a shell object (Config.docx).

This approach was a bit convoluted and has been replaced by a more obvious
solution utilising the registry.

A new user specific application key 'Software\\Linn\\LitePipeTestApp\\UDN'
has been created to hold the UDN string.

It is read on startup. If not found it is created with the default value
"ExampleDevice".

Tested on Windows 7 and 8.