Store permanent configuration properties in the Windows registry.
This commit adds  a new class, ConfigRegStore, which implements the
IStoreReadWrite interface for persistent storage. This replaces the RAM
based example previously passed to the MediaPlayer.

The registry properties are stored under the "Software\\Linn\\LitePipeTestApp"
key.

In addition the read-only property store, RamStore, has been added to the
project to allow property customisation.

The previous example code storing the UDN in a string registry key is
redundant and has been removed.