Fix message posting to UI thread and tidy up.
This commit is mainly a standardisation of coding style:
  - Naming conventions
  - Brace placement
  - Comment clarification.

In addition a bug was fixed relating to the 'args' parameter passed
during Media Player thread creation from the UI thread.

The scope of this variable was incorrect resulting in the data
received by InitAndRunMediaPlayer() being garbage.

The argument data is now stored globally keeping it valid for the
lifetime of the application.