Flush audio buffers when we switch input
When we change input to a source which cannot share the pipeline with other sources,
we must flush outstanding buffers.
Note that on OSX this will also reset DSP state, so there could be a small audio glitch
on restart as the DSP is set up again.  We will ramp the audio up in ohMediaPlayer
on restart so hopefully this should not be very noticeable.