Raspian - Add PresentationUI
This commit adds a Web Application UI. Based on the implementation in
LitePipe::TestMediaPlayer.cpp

In the absence of an installer it is required that the folder containing
the application also contains the 'res' folder obtained from the
dependencies. (Linux-armhf/ohMediaPlayer/res)

Once an installer is available the 'res' folder will be automatically
installed alongside the user specific persistent storage file.