Win32 - Add PresentationUI resources to the installer
This commit adds the resource folder, dependencies/Windows-x86/ohMediaPlayer/res
to the installation package.

The folder is installed in the same folder as the application.

The installer script has been modified to install a release build unless
otherwise instructed.

1. Install Debug build
   iscc /dDebug /dMySrcDir="C:\Users\alzinho\workspaces\TestAppsTidied" LitePipeSampleAppInstaller.iss

2. Install Release build
   iscc /dMySrcDir="C:\Users\alzinho\workspaces\TestAppsTidied" LitePipeSampleAppInstaller.iss

Additionally the libraries required for the PresentationUI have been added
to the project file for release builds as well as debug.