Hook up user notification for updates.
On OSX we will check for updates and then notify the user via NSUserNotificationCenter.
This will add a standard notification to the notifiation panel which can then be clicked to
kick off an update.
We then pop up a dialog confirming whether the user wants to download an update.
If they confirm, we launch the default browser with the uri of the updated app.