Raspian - Add Debian package generation capability
This commit adds a script to generate a Debian installation script
for the application.

The script uses the 'fpm' and ' dpkg' utilitiies.

FPM is ganeral package generator which can generate packages of multiple
formats.

it is installed as follows:

  - sudo apt-get install ruby-dev
  - sudo gem install fpm

The application binary is installed to '/bin' and the icon and resources
are installed to '/bin/share'.

An example script usage would be:
  GeneratePkg.pl -a=litepipe-test-app -v=1.0.0.0

This generated the following application package:
  litepipe-test-app

There are some caveats with the generated package:
  - There is no copyright file
  - The changelog is the default.

These items can be filled in once we know what license/copyright etc. will
be used.