Give raspian builds a unique UDN / name per host
We want to be able to run multiple builds for each platform concurrently on different devices.
We would also like to be able to identify the Room name of each device for easier device selection.

Change the UDN and room setup to use the hostname of the device.
Also add the 'LINN' hex string to the start of the UDN to let Konfig
recognise it as a configurable build.