Win32 - Fix crash on subnet change from UI.
The application crashes when a new subnet is selected from the
UI.

This action causes the MediaPlayer to be restarted with the new
subnet.

The fix is to remove the 'static' from the declaration of 'room'
in InitAndRunMediaPlayer() as on a subsequent call to this function
the previous setting will be invalid.