Win32 - Fix apparent UI hang on audio session disconnection.
While testing Foobar 2000 with the WASAPI Exclusive Mode plugin it was noted
that the LitePipeTestApp would appear to hang when it's audio session
was terminated while playing (due to exclusive mode session being given
preference).

On session termination due to the above or the sound card being reconfigured
the Media Player is stopped and the user is prompted, via a system tray
notification, to restart the application.

Since the initial implementation the application menu has been updated
to contain information from the media player such as available subnets
and playback controls. For obvious reasons this information is
not available when the media player isn't running.

The media player state is now taken into account when generating the
application menu and unavailable items are visible but unselectable.