Reduce audio buffering to a minimum and refactor host audio class
We want to start playing as soon as we receive samples to get the best
songcast sync behaviour.  Reduce the buffers to 250 samples (doing less
than that causes audio to fail to render) and migrate all the (now simpler)
host audio code into the Animator class itself.

Make debug printing optional but leave it in as it is useful.