Win32 - Resample streams to system mix format.
This commit adds support for stream resampling to allow streams with
sample rates/bit depths other than that currently configured for the soundcard
to play.

To accomplish this code from the 'bitsPerSampleConv2i::WWMFResamplerTest'
project has been plumbed in to execute the translation.

Project URL:
https://code.google.com/p/bitspersampleconv2/source/browse/#svn%2Ftrunk%2FWWMFResamplerTest

The resampler takes care of:
  - mono -> stereo conversion.
  - bit depth conversion
  - sample rate conversion.

This allows the ProcessorPcmWASAPI to be simplified, by removing the
mono->stereo conversion code.

If the stream cannot be played the resampler is configured to translate
to the current system mix. Data produced by the PCM processor is passed
through the resampler before being placed in the native audio render buffer.

It would have been nice to have the translation done in the PCM processor,
but this didn't work well in the 'sample' processing functions.