Win32 - Fetch applicaiton updates for manual install
This commit updates the application to fetch updates when the user
clicks 'Update' item in the context menu.

The 'fetch' is done by opening the update URI in the default browser
then exiting the application.

When the downloaded update is executed the application will be installed
and run.

The installer script has been updated to prevent the user modifying the
application install location and start menu entry on an update.