Linux - Create Self Extracting Installer
This commit adds the ability to create a self extracting package installer.

This functionality is targeted at platforms which do not offer
a method of installing a Debian package while resolving missing dependencies.

The generated installer performs the following actions:
  1. Extract the embedded package to a temporary folder.
  2. Use 'dpkg' to install the package
  3. Use 'apt-get -f' to 'fix' any broken dependencies.

A new command line option, '--installer', has been added to the release
generation script.

With this option enabled the installer is constructed as follows:
  1. Create the installer from the template bash script
     'Installer-Template.txt'.

     The marker text '<PUT PKG FILE HERE>' is replaced with the release
     package name.
  2. The release package is then appended to the installer.
  3. The installer is made executable.