Linua (Ubuntu) - Optional feature setup code
This commit adds sample setup code for the following optional features:

  - Qobuz
  - Tidal
  - Radio
  - AAC codec
  - MP3 codec

Each feature is guarded be an "ENABLE_*" define which, by default, is not
defined.

The implementor is left to supply tokens for services which require them.

The Ubuntu makefile has been updated to accept an extra set of command line
options which will enable the above features and link against any
extra required libraries.

The options are:

  - ENABLE_QOBUZ=0
  - ENABLE_TIDAL=0
  - ENABLE_RADIO=0
  - ENABLE_AAC=0
  - ENABLE_MP3=0