Linux (Raspbian/Ubuntu) - Fix track advance failure
This commit fixes an issue with track advance whereby the subsequent
track would not play.

This issue is down to the snd_pcm_drain() operation performed in
ProcessDrain().

The PCM must be 'prepared' (snd_pcm_prepare()) after the drain
before it will accept more PCM data.

This operation was missing. Now it isn't.