Make menu icons work with dock modes (light and dark)
NSImage has an attribute which allows you to specify that the image should
be used as a template. This requires that the images contain only black pixels
(but can have translucency).
Change our menu icons to be black only, and add the 'template' flag to the
loaded icon images. This allows us to display correctly on Yosemite light
and dark modes.