Osx - Introduce CP proxies for each audio source
This commit introduces control point proxies for each audio source:
  - Playlist
  - Radio
  - Receiver (Songcast)
  - UpnpAV

A proxy is also created for the 'Product' service. This proxy is used to
obtain asynchronous notifications of an audio source change, via the
source index changed callback.

This callback will read and parse the source Xml to identify the source
at the index.

The individual source proxies listen for and log transport state changes,
however there a no asynchronous messages posted to the UI. The transport
state is queried by the IU code, via the can*() API's each time the menu
is displayed.

The ControlPoint proxy keeps track of the active source, and passes UI
playback actions to the correct proxy.

NB. The UpnpAV proxy does not have an asynchronous method for tracking
    transport changes. For this proxy the pipeline state change notification
    is used.