Osx - Control native audio from ControlPointProxy
This commit adds control of the native audio from the individual control
point proxies.

This allows the native audio state to be kept in sync with the current
transport/pipeline state.

Passing a reference to the Osx driver to the ControlPointPorxy allows it
call the driver pause() and resume() functions as required.