Osx - UpdateChecker now gets version from Info.plist
On the Osx platform the application version used by the system is stored
in the Info.plist file.

Prior to this commit the update check code used the CURRENT_VERSION define
from version.h to determine the application version.

To simplify the release process the version used for the update check is
now also read from the Info.plist file avoiding having to set the version
in multiple places..