Linux (Raspbian/Debian) - External Codecs
This commit adds a native, libav based, codec to handle MP3/AAC.

The use of the codec is enabled at build time by adding 'USE_LIBAVCODEC=0'
to the build line. Support for the individual formats is still controlled
by the defines in OptionalFeatures.h

There are issues with seeking within streams which will be investigated
in a follow up review.

The Pi build isn't behaving well. MP3 wont play, AAC does but sounds awful.
That'll need to be sorted in a follow up.

NB. The version of libav on Ubuntu 14.04 decode AAC to FLTP and MP3 to S16P.
    This may not be the case for all versions of libav.

A number of small buffer leaks in the GTK code have also been plugged.