Linux (Raspbian) - Fix MP3 Playback
This commit fixes an issue on Raspbian whereby the application would crash
when attempting to play an Mp3 track without first adding it to the
playlist.

The problem was iController->Read() was throwing a 'CodecStreamStopped'
exception which wasn't being caught.

This is now handled in avCodecRead() as per the 'CodecStreamEnded'
exception.