Linux - More Seek Tweaks
This commit updates to the latest dependencies and tweak the seek logic.

The latest dependencies have had a positive effect on the seek
behaviour.

However for MP3 it was noted that performing a frame read after
a seek attempt still did not force execution of out seek callback.

In this instance we assume the packet read is valid and call TrySeekTo()
with it's offset in the stream.