Handle over-length room and product names.
The app was using the machine name as Room name, and wasn't aware that the
Product class uses fixed buffer sizes for these attributes. Passing a buffer
with more bytes than it can hold internally causes a crash (eg if your
machine name was more than 20 bytes long).
Add some code to handle this by ensuring that the buffers passed to the
MediaPlayer constructor are always bounded by the maxbytes constants declared
in Product.h.