Fix OSX suspend part of bug #2
On OSX we would go into sleep mode, but never re-appear on the network once
the system had resumed.

Register the app for wake/sleep notifications and puause resume the ohNet stack
and Audioqueue as appropriate.

Note that we don't currently re-sync to sender if we're in Soncast receiver
mode when we sleep; we play through existing caches. These should really be
invalidated on resume.