Linux - Add (Optional) Command Line Option To Set Subnet
This commit adds a command line option to specify the subnet to join.

If this is not available the application will revert to the last used
subnet.

New Usage: openhome-player [subnet address]

The use case for this option is primarily for headless builds running on
platforms with multiple network interfaces.