Win32 - Extend the Win32 PCM Handers To Convert To Native The Bit Depth
This commit extends the functionality of the 32 bit PCM handlers to handle
the PCM generated by the ramper.

No matter the PCM format being output by the current codec the pipeline
ramper element produces 32 bit PCM.

This commit converts the bit depth to that produced by the current codec
to make it compatible with the current audio subsystem configuration.