Linux - Extend the Linux PCM Handers To Convert To The Native Bit Depth
This commit extends the functionality of the 32 bit PCM handlers to handle
the PCM generated by the ramper.

The PcmProcessorLe processor always operates with ALSA configured for 16 bit
PCM. Thus the 32 bit handlers already convert to 16 bit. Nothing else needs
to be done.

The PcmProcessorLe32 processor operates with ALSA configured for 32 bit PCM
for 24/32 bit streams and ALSA configured for 16 but PCM for 8/16 bit
streams.

For the later case we update the 32 bit PCM handlers to convert to 16 bit
PCM.