Osx - Latest Dependencies Fix Some Osx Hangs
This commit fixes most issues resulting in the Osx player hanging.

The player was seen to hang:

1. On source change.
2. On exit when no audio is playing.

This issue relates to the control point proxy reacting to a pipeline/transport
state change to pause/stopped and pausing the audio queue.

With the audio blocking further pipeline message reads the pipeline never
regains buffering/playing state and remains blocked causing the hang.

Moving the 'pause' from the proxy to the MsgHalt handler and allowing the proxy
to 'resume' the audio queue on 'Buffering' in 'Addition' to 'Playing'
resolves the above issues.

There remains an issue whereby the player can hang when starting to play
a radio station ...