Osx - Latest Dependencies Fix Pipeline Animator Hang
Currently the native AudioQueue is paused on receipt of a MsgHalt. This
message is received after all available MsgPlayable messages have been
consumed thus avoiding the audio queue buffer fill blocking pipeline
animation (as the audio queue will stop requesting data when stopped).

The resumption of the AudioQueue happens when the pipeline/transport state
changes to buffering/playing. This approach works well in the main, but is
problematic when changing from one radio channel to another channel of the
same audio format. MsgHalts are received without any corresponding
pipeline/strasnport state change to restart the AudioQueue.

The desired behaviour is that the pipeline animation blocks/unblocks naturally
with the frequency of MsgPlayable pulls regulated by AudioQueue.

This commit moves the resumption of the AudioQueue from the Control Point Proxy
to the MsgPlayable handler, thus giving us our desired behaviour.