Linux - Fix Crash On Play Of Unplayable Audio File
The native audio device on the Raspberry Pi doesn't support high sample
rates.

Prior to this commit when this situation arose the driver would read
and discard the track audio data. This approach now causes a crash as it
drains the pipeline too quickly.

The solution is to update PipelineAnimatorDelayJiffies() to query Alsa
with regards to the sample rate and reject the track, by throwing a
SampleRateUnsupported exception, if the sample rate is not supported.