Osx - Latest Dependencies Fix Audio Glitch
The previous commit to address an application hang due to an attempt to
split a MsgPlayable() on a non-frame boundary inadvertently caused an audio
glitch on streams of a high sample rate.

The simple approach of queuing the current audio queue buffer as is and
deferring any MsgPlayable that didn't fit in the audio buffer in it's
entirety until the next run was to blame.

This commit reverts to the previous approach of splitting the MsgPlayable
to fill the audio queue buffer, with the proviso that the split occurs
on a frame boundary.

This is done by ensuring the split occurs on the lowest common multiple
of the available sample sizes (1/2/3/4 bytes) multiplied by the maximum
number of channels per frame.