Osx - Improve Konfig Support.
This commit updates the presentation Url handling in PresentationUrlChanged()
to bring it in line with recent dependency updates.

This allows the Linn Konfig utility to display the Web UI Configuration
page for the player.

To enable Konfig to display the model icon for the player:

  - Ramstore is constructed with an image filename.
    This is used as the kBufModelImageUrl value.

    NB. Static options that do not start with 'http' will result in calls
        to ExampleMediaPLayer::WriteResource() upon request.

  - IconOpenHome.h has been added to the code base.
    This contains the icon as a byte array.

  - WriteResource() has been updated to use the resource writer
    to pass the icon data to the requesting control point.

NB. Konfig is not displaying the Web Config UI for some, as yet,  unknown
    reason